A FOSS implementation of Delta
