+++ app/src/foss/java/website/leifs/delta/activities/MainActivity.java
@@ -1,0 +1,14 @@
+package website.leifs.delta.activities;
+
+import androidx.annotation.NonNull;
+
+import candybar.lib.activities.CandyBarMainActivity;
+
+public class MainActivity extends CandyBarMainActivity {
+
+    @NonNull
+    @Override
+    public ActivityConfiguration onInit() {
+        return new ActivityConfiguration();
+    }
+}
