+++ app/src/main/res/values/dashboard_icon_pack.xml
@@ -3,29 +3,4 @@
     xmlns:tools="http://schemas.android.com/tools"
     tools:ignore="MissingTranslation">
 
-    <!-- YOUR ICON PACK NAME -->
-    <string name="app_name">Delta</string>
-
-    <string name="theme_title">Delta</string>
-    <string name="theme_name">Delta</string>
-
-    <string name="theme_author">Leif Niemczik</string>
-    <string name="developer_name">Leif Niemczik</string>
-    <string name="developerName">Leif Niemczik</string>
-    <string name="authorName">Leif Niemczik</string>
-
-    <string name="authorLink">https://leifs.website</string>
-    <string name="developer_link">https://leifs.website/</string>
-
-    <string name="theme_description">Pastel Goodness</string>
-    <string name="theme_info">Pastel Goodness</string>
-    <string name="theme_ad_msg">Pastel Goodness</string>
-
-    <string name="theme_feature">Pastel Goodness</string>
-
-    <string name="theme_preview1">preview1</string>
-    <string name="theme_preview2">preview2</string>
-
-    <string name="iconpack">website.leifs.delta</string>
-
 </resources>
