+++ app/src/play/java/website/leifs/delta/activities/MainActivity.java
@@ -1,0 +1,20 @@
+package website.leifs.delta.activities;
+
+import androidx.annotation.NonNull;
+
+import candybar.lib.activities.CandyBarMainActivity;
+import website.leifs.delta.licenses.License;
+
+public class MainActivity extends CandyBarMainActivity {
+
+    @NonNull
+    @Override
+    public ActivityConfiguration onInit() {
+        return new ActivityConfiguration()
+                .setLicenseCheckerEnabled(License.isLicenseCheckerEnabled())
+                .setLicenseKey(License.getLicenseKey())
+                .setRandomString(License.getRandomString())
+                .setDonationProductsId(License.getDonationProductsId())
+                .setPremiumRequestProducts(License.getPremiumRequestProductsId(), License.getPremiumRequestProductsCount());
+    }
+}
