+++ app/src/play/java/website/leifs/delta/licenses/License.java
